Well now...   MORE quests!
ENJOY!

made a number of change around the side-quests.  mainly added several more.   need to figure out a more maintainble way to handle sidequest completion in the dequeuue.   prototyped a reward section in the templates.  poor-bobly will want to add a failure secton too and figure out how to make the deque generalized off the templates.. or update the dosidequest function(s)... but it seems to be mostly working!  see a few text/replacement issues but otherwise working  :)